requirements.txt
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-webpack-loader==0.6.0
docxtpl==0.9.0